third_party/blink/web_tests/fullscreen/video-specified-size-expected.txt

EVENT(webkitfullscreenchange) TEST(video.clientWidth==document.body.clientWidth) OK
END OF TEST
